[skip ci] NCN-126 expansion and improvement of plugin api

+++ common/indicators-api/src/calculation.rs
@@ -1,11 +1,11 @@
 use anyhow::{bail, Error};
 
-pub fn moving_average(values: &[f64], length: usize) -> Result<Vec<f64>, Error> {
-    if values.is_empty() || values.len() < length {
+pub fn moving_average(values: &[f64], length: u16) -> Result<Vec<f64>, Error> {
+    if values.is_empty() || values.len() < length as usize {
         bail!("Moving average length too long for this set of values.")
     }
     let mut tail_pointer = 0usize;
-    let mut head_pointer = length;
+    let mut head_pointer = length as usize;
     let mut result = Vec::new();
     while head_pointer <= values.len() {
         let range_avg = average(&values[tail_pointer..head_pointer]);
@@ -28,8 +28,7 @@ mod tests {
     #[test]
     fn test_moving_average() {
         let values = [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0];
-        let result = moving_average(&values, 7)
-            .expect("Error during moving average calculation");
+        let result = moving_average(&values, 7).expect("Error during moving average calculation");
         assert_eq!(result.len(), 5);
         assert_eq!(result.first().unwrap(), &3.0);
         assert_eq!(result.get(1).unwrap(), &4.0);
